extracted event_handler
