+++ handlers/help_viewer.py
@@ -3,7 +3,7 @@ from handlers.main_game_event_handler import MainGameEventHandler
 import keys
 import strings
 from engine import Engine
-from handlers.input_handlers import EventHandler
+from handlers.event_handler import EventHandler
 from typing import Optional
