+++ handlers/inspect_item_handler.py
@@ -1,7 +1,7 @@
 from engine import Engine
 from entity import Item
 from handlers.inventory_event_handler import InventoryEventHandler
-from handlers.input_handlers import EventHandler
+from handlers.event_handler import EventHandler
 from handlers.input_handlers import ActionOrHandler
 from typing import Optional
