+++ handlers/main_game_event_handler.py
@@ -4,7 +4,7 @@ import keys
 from actions.bump_action import BumpAction
 from actions.pickup_action import PickupAction
 from actions.wait_action import WaitAction
-from handlers.input_handlers import EventHandler
+from handlers.event_handler import EventHandler
 from handlers.input_handlers import ActionOrHandler
 from typing import Optional
