+++ handlers/pause_viewer.py
@@ -2,7 +2,7 @@ import color
 import exceptions
 from handlers.main_game_event_handler import MainGameEventHandler
 import keys
-from handlers.input_handlers import EventHandler
+from handlers.event_handler import EventHandler
 from typing import Optional
