+++ main.py
@@ -3,6 +3,7 @@ import traceback
 
 import tcod
 
+import handlers.event_handler
 import handlers.main_game_event_handler
 from helpers import resource_path
 import color
@@ -14,7 +15,7 @@ import sizes
 
 def save_game(handler: input_handlers.BaseEventHandler) -> None:
     """If the current event handler has an active Engine then save it."""
-    if isinstance(handler, input_handlers.EventHandler):
+    if isinstance(handler, handlers.event_handler.EventHandler):
         handler.engine.save_as(handler.engine.save_path)
         print("Game saved.")
 
@@ -53,7 +54,7 @@ def main() -> None:
                 except Exception:  # Handle exceptions in game.
                     traceback.print_exc()  # Print error to stderr.
                     # Then print the error to the message log.
-                    if isinstance(handler, input_handlers.EventHandler):
+                    if isinstance(handler, handlers.event_handler.EventHandler):
                         handler.engine.message_log.add_message(traceback.format_exc(), color.error)
         except exceptions.QuitWithoutSaving:  # Quit to desktop
             raise
